refactor(EndedJanggiGameManager): 디렉터리 변경
